Move product categories related data from types to constants
